Refactored to remove unnecessary flexibility
The flexibility afforded by using ChannelBehavior, RgbDeviceBehavior,
etc. was unnecessary. This device is primarily intended for controlling
RGB devices, and I feel I shouldn't be wasting so much time trying to
implement flexibility which allows channels to be used for various
devices and other crazy things.

This makes developing for the RGB controller less overwhelming, as I
don't need to build in as much flexibility at every level.